backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.23
pydantic>=1.8.2
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.5
pytest>=6.2.5
requests>=2.26.0
aiohttp>=3.8.1
python-dotenv>=0.19.0
bcrypt>=3.2.0 